+++ test/ReactViews/Tools/ItemSearchTool/ItemSearchToolSpec.tsx
@@ -71,7 +71,7 @@ describe("ItemSearchTool", function() {
     item = new TestSearchableItem("test", terria);
     item.setTrait(CommonStrata.user, "search", {
       providerType: "testProvider",
-      options: {},
+      providerOptions: {},
       resultTemplate: undefined,
       highlightColor: undefined,
       parameters: []